Update scripts, dotfiles and Brewfile

+++ install.sh
@@ -2,14 +2,15 @@
 
 set -eu
 
-export DOTFILES_ROOT="$HOME/dotfiles"
+readonly DOTFILES_ROOT="$HOME/dotfiles"
+readonly DOTFILES="dotfiles"
 
 #===========================================================================#
 # common functions
 #===========================================================================#
 
 print() {
-  printf "\e[4mdotfiles\e[0m: "
+  printf "\e[4m$DOTFILES\e[0m: "
   echo "$@"
 }
 
@@ -25,15 +26,28 @@ command_exist() {
 
 #===========================================================================#
 
-if [[ -d "$DOTFILES_ROOT" ]]; then
-  print "Dotfiles is already installed. Override it."
-  cmd_exec rm -fr $DOTFILES_ROOT
-  print
-fi
-
 if ! command_exist git; then
   print "Error: Dotfiles need git."
   exit 1
+fi
+
+if [[ -d "$DOTFILES_ROOT" ]]; then
+  print "Dotfiles is already installed."
+
+  while true; do
+    print -n "Do you want to overwrite it? [Y/n] "
+    read answer
+    case "$answer" in
+      [yY])
+        cmd_exec rm -fr $DOTFILES_ROOT
+        print
+        break
+        ;;
+      [nN])
+        exit 1
+        ;;
+    esac
+  done
 fi
 
 cmd_exec git clone --recursive https://github.com/koki-sato/dotfiles.git "$DOTFILES_ROOT"
